% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{boot.silhouette}
\alias{boot.silhouette}
\title{Produces silhouette plots}
\usage{
boot.silhouette(clusboot.out, ...)
}
\arguments{
\item{clusboot.out}{an object of class clusboot}

\item{...}{more arguments to be passed to \code{barplot()}}
}
\value{
list of silhouette widths
}
\description{
Produces silhouette plots
}
\examples{
out <- clusboot (scale(case.study.psychiatrist), B=100, k=6, clustering.func=complete.linkage)
boot.silhouette(out)
}
