% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{calc.silhouette}
\alias{calc.silhouette}
\title{Computes the silhouette vales based on proportion of times items cluster together}
\usage{
calc.silhouette(clusboot.out)
}
\arguments{
\item{clusboot.out}{an object of class clusboot}
}
\value{
an object of class clusboot
}
\description{
Computes the silhouette vales based on proportion of times items cluster together
}
