\name{ClusBoot-package}
\alias{ClusBoot-package}
\alias{ClusBoot}
\docType{package}
\title{
Performs bootstrap on a cluster analysis output
}
\description{
Any clustering output is subjected to a bootstrap procedure to determine the stability of the clustering solution. The results are displayed
in the form of a Multi-dimensional scaling plot and a silhouette plot.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ClusBoot}
\packageIndices{ClusBoot}
A package to perform bootstrap on any cluster analysis output and visualise the results.
For more infomation see \url{https://slubbe.wixsite.com/academic-cv/conference-presentations}
}
\author{
Sugnet Lubbe slubbe@sun.ac.za

Maintainer: Sugnet Lubbe slubbe@sun.ac.za
}
\references{
Lubbe, S. Visualisations associated with bootstrapping cluster analysis. Data Science, Statistics and Visualisation conference, July 2017, Lisbon, Portugal.
}
