% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{boot.silhouette}
\alias{boot.silhouette}
\title{Constructs a silhouette plot based on proportion of times items cluster together}
\usage{
boot.silhouette(clusboot.out, ...)
}
\arguments{
\item{clusboot.out}{an object of class clusboot}

\item{...}{more arguments to be passed to \code{plot()}}
}
\value{
vector of silhouette widths for each of the clusters
}
\description{
Constructs a silhouette plot based on proportion of times items cluster together
}
\examples{
out <- clusboot (scale(case.study.psychiatrist), B=100, k=6, clustering.func=complete.linkage)
boot.silhouette(out)
}
