% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{clusboot}
\alias{clusboot}
\title{Performs bootstrap on a cluster analysis output}
\usage{
clusboot(datmat, B = 1000, clustering.func = complete.linkage, ...)
}
\arguments{
\item{datmat}{a data matrix or distance object which will be the input to the clustering function}

\item{B}{number of bootstrap replicates}

\item{clustering.func}{the function which will perform the clustering and output a vector of cluster memberships}

\item{...}{more arguments to be passed to the clustering function}
}
\value{
an object of type clusboot
}
\description{
Performs bootstrap on a cluster analysis output
}
\examples{
clusboot (scale(case.study.psychiatrist), B=100, k=6, clustering.func=complete.linkage)
}
