% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{complete.linkage}
\alias{complete.linkage}
\title{Wrapper function for performing complete linkage clustering}
\usage{
complete.linkage(X, k)
}
\arguments{
\item{X}{samples x variables data matrix}

\item{k}{number of clusters}
}
\value{
vector of cluster memberships
}
\description{
Wrapper function for performing complete linkage clustering
}
\examples{
complete.linkage(scale(case.study.psychiatrist), k=6)
}
